fixbug: fixed hot bug TelegramUser.user_id.

+++ backend/telegram/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 class TelegramUser(models.Model):
     # fields copied from https://github.com/itisnotyourenv/django-telegram-bot/blob/main/utils/models.py
-    user_id = models.IntegerField(primary_key=True)
+    user_id = models.PositiveBigIntegerField(primary_key=True)
     username = models.CharField(max_length=32, null=True, blank=True)
     first_name = models.CharField(max_length=256)
     last_name = models.CharField(max_length=256, null=True, blank=True)
